+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add a plugin to the demo from any git repository, with the matching version pre-selected
 - Support for offline mode
+- Support for custom poster tool card
+- Integration of CAMP plugin registry
 
 ### Changed